pyproject.toml
[tool.poetry]
name = "discord-diablo4-notify-py"
version = "0.1.0"
description = ""
authors = ["Dylan <40823034+linty1997@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "discord_diablo4_notify_py"}]

[tool.poetry.dependencies]
python = "^3.11"
py-cord = "^2.4.1"
python-dotenv = "^1.0.0"
workers-kv-py = "^1.2.2"
flatten-dict = "^0.4.2"
beautifulsoup4 = "^4.12.2"
playwright = "^1.34.0"
tweepy = "^4.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
